+++ src/lib.rs
@@ -19,5 +19,4 @@ mod tests;
 pub use crate::context::{Context, ContextHolidays};
 pub use crate::opening_hours::OpeningHours;
 pub use crate::utils::range::DateTimeRange;
-pub use opening_hours_syntax::error::Error as ParserError;
 pub use opening_hours_syntax::rules::RuleKind;